requirements.txt
annotated-types==0.6.0
anyio==3.7.1
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
exceptiongroup==1.2.0
fastapi==0.104.1
h11==0.14.0
idna==3.6
mysql-connector==2.2.9
mysql-connector-python==8.2.0
protobuf==4.21.12
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyMySQL==1.1.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
uvicorn==0.24.0.post1
